Create a component for the entity stat overview section
